+++ src/lib/posts.js
@@ -1,0 +1,4 @@
+/* export const addPost=(text)=>{
+    // conectar a firestore (addDoc o setDoc)
+    // guardar el post en la coleccion posts
+} */
